+++ examples/ditributed_map.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel
-from indexify import indexify_function, indexify_router, Graph
+from indexify import indexify_function, indexify_router, GraphDS
 from typing import List, Union
 
 @indexify_function()
@@ -12,7 +12,7 @@ def squared(x: int) -> int:
 
 if __name__ == '__main__':
     from indexify import create_client
-    g = Graph(name="map_square", start_node=generate_sequence, description="Simple Sequence Summer")
+    g = GraphDS(name="map_square", start_node=generate_sequence, description="Simple Sequence Summer")
     g.add_edge(generate_sequence, squared)
 
     client = create_client(service_url="http://100.106.216.46:8900")
