+++ examples/readme/distributed_map.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel
-from indexify import indexify_function, indexify_router, Graph
+from indexify import indexify_function, indexify_router, GraphDS
 from typing import List, Union
 
 @indexify_function()
@@ -11,7 +11,7 @@ def squared(x: int) -> int:
     return x * x
 
 if __name__ == '__main__':
-    g = Graph(name="sequence_summer", start_node=generate_sequence, description="Simple Sequence Summer")
+    g = GraphDS(name="sequence_summer", start_node=generate_sequence, description="Simple Sequence Summer")
     g.add_edge(generate_sequence, squared)
 
     from indexify import create_client
