+++ examples/readme/readme_example.py
@@ -1,8 +1,8 @@
 from pydantic import BaseModel
-from indexify import indexify_function, Graph, indexify_router
+from indexify import indexify_function, indexify_router
 from typing import List, Union
 
-from indexify.functions_sdk.sup_graph import SupGraph
+from indexify.functions_sdk.graph import Graph
 
 
 class Total(BaseModel):
@@ -33,7 +33,7 @@ def dynamic_router(val: Total) -> List[Union[square, cube]]:
 
 
 def run_local(graph_name, a, fn):
-    g = SupGraph(
+    g = Graph(
         name=graph_name,
         start_node=generate_numbers,
         description="Simple Sequence Summer",
@@ -48,7 +48,7 @@ def run_local(graph_name, a, fn):
 
 
 def run_remote(graph_name, a, fn):
-    g = SupGraph(
+    g = Graph(
         name=graph_name,
         start_node=generate_numbers,
         description="Simple Sequence Summer",
@@ -63,7 +63,7 @@ def run_remote(graph_name, a, fn):
     print(result)
 
 def register_remote_graph(graph_name):
-    g = SupGraph(
+    g = Graph(
         name=graph_name,
         start_node=generate_numbers,
         description="Simple Sequence Summer",
@@ -90,8 +90,8 @@ if __name__ == '__main__':
     ### Test the graph with a remote run, make sure to register it first
     # graph_name = "sequence-summer"
     # register_remote_graph(graph_name=graph_name)
-    #
-    # g = SupGraph.from_server(server_url="http://localhost:8900", namespace="default", name=graph_name)
+
+    # g = Graph.from_server(server_url="http://localhost:8900", namespace="default", name=graph_name)
     # invocation_id = g.run(a=6, block_until_done=True)
     # result = g.graph_outputs(invocation_id, "square")
     # print(result)
