+++ python-sdk/indexify/__init__.py
@@ -1,6 +1,6 @@
 from . import data_loaders
 from .client import create_client
-from .functions_sdk.graph import Graph
+from .functions_sdk.graphds import GraphDS
 from .functions_sdk.image import Image
 from .functions_sdk.indexify_functions import (
     indexify_function,
@@ -12,7 +12,7 @@ from .settings import DEFAULT_SERVICE_URL
 
 __all__ = [
     "data_loaders",
-    "Graph",
+    "GraphDS",
     "Image",
     "indexify_function",
     "indexify_router",
