+++ python-sdk/tests/test_graph_update.py
@@ -4,7 +4,7 @@ import unittest
 from pydantic import BaseModel
 
 from indexify import create_client
-from indexify.functions_sdk.graph import Graph
+from indexify.functions_sdk.graphds import GraphDS
 from indexify.functions_sdk.indexify_functions import indexify_function
 
 
@@ -24,7 +24,7 @@ def update2(x: Object) -> Object:
 
 class TestGraphUpdate(unittest.TestCase):
     def test_graph_update(self):
-        g = Graph(name="updategraph1", start_node=update)
+        g = GraphDS(name="updategraph1", start_node=update)
         client = create_client()
         client.register_compute_graph(g)
         invocation_id = client.invoke_graph_with_object(
@@ -33,7 +33,7 @@ class TestGraphUpdate(unittest.TestCase):
         output = client.graph_outputs(g.name, invocation_id, fn_name="update")
         self.assertEqual(output[0], Object(x="ab"))
 
-        g = Graph(name="updategraph1", start_node=update2)
+        g = GraphDS(name="updategraph1", start_node=update2)
         client.register_compute_graph(g)
         client.rerun_graph(g.name)
         time.sleep(1)
